Move Switch files to another folder
